webspaced: Split webspace and manager code
